feat(test): workaround for issues with IPv6/IPv4 dual-stacking
The testing library (supertest) will start the application on a randomly allocated ("ephemeral")
port on localhost, and Node will bind to the IPv6 localhost ("::") by default if it is
available.

Supertest then uses a hard-coded hostname in IPv4 format ("127.0.0.1"). Normally operating
systems support a feature called "dual-stacking" whereby the IPv4 localhost address is
automatically translated to the IPv6 host, and apps trying to connect using IPv4 should be
able to connect to the IPv6 socket.

If this feature is unavailable or not able to be used for some reason, the test suite will
fail to connect to the local app instance and will fail with a timeout.

If the INTAKE24_TEST_SET_HOST environment variable is defined, the test suite will start the
application using the port and hostname configured in the environment file instead of relying
on supertest defaults. It is possible to pass a server reference instead of the app reference
to supertest functions to achieve this.

By setting the hostname to the IPv4 localhost ("127.0.0.1") in the environment file the
application will bind to an IPv4 socket thus avoiding the potential IPv6/IPv4 mismatch.