feat: Drink sliding scales rework
This commit includes the following new features and changes:

Currently sliding scale data for drinks is stored in a legacy format that has multiple issues:

* It relies on a pre-rendered overlay image to render the fill level image, which is
  a compatibility workaround that is no longer necessary,

* It addresses the image resources directly circumventing the image processing system
  thus violating the configurable image dimensions and the source image storage
  conventions,

* It does not store the vector data from the original image outline, making editing the data
  or changing the way it is displayed impossible,

* It does not record the original volume measurements data, only the normalised samples
  remapped to [0, 1] range, which makes amending the measurements data impossible.

This commit introduces a new drink scale data format that resolves all of the above issues:

* The sliding scale is now rendered client-side using an SVG overlay in line with the rest
  of the system (via DrinkScaleV2Panel component),

* Drink scale images are processed and resized using the standard image processing pipeline
  in line with the rest of the system,

* Drink scale outline is now stored in the vector form,

* Original measurements (in mm) are recorded alongside the normalised values.

Since the previous format is destructive it will have to exist alongside the new format until
the sliding scale data can be recovered and converted into the new format or recreated.

The current API will send new sliding scale data when it is available and old data when it
is not, and the client-side UI will choose which component to use to render it based on the
version tag.

The legacy route should be deleted once the drink scale data is converted.

A new CLI interface command is introduced, 'convert-drink-scale', which can read the legacy
drink scale data (consisting of SVG vector files defining the container selection image maps
and the fillable volume shape and CSV files contaning the volume measurements) and convert
them into the Intake24 package format which can them be imported into the system as usual.

The CLI package importer tool, available via the 'import-package' command, has been updated
to import the drink scale data in the new format.

The new database code (currently limited to drinkware-set.service.ts) is written using Kysely
query builder library in order to evaluate it against Sequelize.